Streamline space routing
This changes space routing to use /space/{id} and space creation
to /space/create for clarity and simplicity. It also changes space
creation to a page rather than a modal, and adds a cancel button for
the creation form.

This is done to simplify development and make urls more understandable.